requirements.txt
asgiref==3.2.7
Django==3.0.7
django-environ==0.4.5
django-storages==1.9.1
djangorestframework==3.11.0
drf-nested-routers==0.91
model-bakery==1.1.0
munch==2.5.0
Pillow==7.2.0
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
